Decouple Rosetta generation and stub generation
