+++ src/annotation/types.d.ts
@@ -1,19 +1,9 @@
-import { BaseReportArgs } from '../base'
+import { BaseAnnotateArgs } from '../base'
 
 /**
  * Arguments for annotation.
  */
-export interface AnnotateArgs extends BaseReportArgs {
-    /**
-     * The directory to write typestub files to.
-     */
-    outputDirectory: string
-
-    /**
-     * The directory to load rosetta files from.
-     */
-    rosetta?: string
-
+export interface AnnotateArgs extends BaseAnnotateArgs {
     /**
      * Whether fields and functions in the generated stubs should be alphabetized.
      */
@@ -25,32 +15,7 @@ export interface AnnotateArgs extends BaseReportArgs {
     includeKahlua: boolean
 
     /**
-     * Whether injection via Rosetta of data that wasn't detected in the source is enabled.
-     */
-    inject: boolean
-
-    /**
      * Whether fields should be treated as strict.
      */
     strictFields: boolean
-
-    /**
-     * Classes which should be excluded from the generated stubs.
-     */
-    exclude?: string[]
-
-    /**
-     * Classes whose fields should be excluded from the generated stubs.
-     */
-    excludeFields?: string[]
-
-    /**
-     * Whether known large definition classes should have their fields excluded.
-     */
-    excludeKnownDefs: boolean
-
-    /**
-     * The format to use for generating Rosetta stubs.
-     */
-    format?: 'json' | 'yml'
 }
