+++ src/base/index.ts
@@ -1,3 +1,4 @@
 export type * from './types'
 export { BaseReader } from './BaseReader'
 export { BaseReporter } from './BaseReporter'
+export { BaseAnnotator } from './BaseAnnotator'
