+++ src/base/types.d.ts
@@ -62,6 +62,41 @@ interface BaseReportArgs extends BaseArgs {
 }
 
 /**
+ * Base arguments for a class that handles annotation.
+ */
+export interface BaseAnnotateArgs extends BaseReportArgs {
+    /**
+     * The directory to write typestub files to.
+     */
+    outputDirectory: string
+
+    /**
+     * The directory to load rosetta files from.
+     */
+    rosetta?: string
+
+    /**
+     * Whether injection via Rosetta of data that wasn't detected in the source is enabled.
+     */
+    inject?: boolean
+
+    /**
+     * Classes which should be excluded from the generated stubs.
+     */
+    exclude?: string[]
+
+    /**
+     * Classes whose fields should be excluded from the generated stubs.
+     */
+    excludeFields?: string[]
+
+    /**
+     * Whether known large definition classes should have their fields excluded.
+     */
+    excludeKnownDefs: boolean
+}
+
+/**
  * Base arguments for Lua readers.
  */
 interface BaseReaderArgs {
