+++ src/main.ts
@@ -5,6 +5,10 @@ import { hideBin } from 'yargs/helpers'
 import { ResolveArgs, Resolver } from './dependency-resolution'
 import { AnalyzeArgs, Analyzer } from './analysis'
 import { AnnotateArgs, Annotator } from './annotation'
+import {
+    RosettaGenerateArgs as GenerateArgs,
+    RosettaGenerator as Generator,
+} from './rosetta'
 
 /**
  * Adds shared yargs options to prefix for all commands.
@@ -62,119 +66,107 @@ const sharedSuffix = (yargs: yargs.Argv) => {
         .wrap(120)
 }
 
+const addExcludeOptions = (yargs: yargs.Argv) => {
+    return yargs
+        .option('exclude', {
+            type: 'array',
+            alias: 'e',
+            string: true,
+            desc: 'Classes to exclude from annotations',
+        })
+        .option('exclude-fields', {
+            type: 'array',
+            string: true,
+            desc: 'Classes to include without fields',
+        })
+        .option('exclude-known-defs', {
+            type: 'boolean',
+            default: true,
+            desc: 'Whether known definition classes should be included without fields',
+        })
+}
+
 /**
  * Adds the CLI options for the annotate command.
  */
 const annotateCommand = (yargs: yargs.Argv) => {
-    return sharedSuffix(
-        sharedPrefix(yargs)
-            .option('output-directory', {
-                type: 'string',
-                alias: 'o',
-                required: true,
-                desc: 'The directory for output stubs',
-            })
-            .option('alphabetize', {
-                type: 'boolean',
-                default: true,
-                desc: 'Whether fields and functions should be alphabetically sorted',
-            })
-            .option('include-kahlua', {
-                type: 'boolean',
-                alias: 'k',
-                desc: 'Whether to generate the kahlua stub.',
-            })
-            .option('inject', {
-                type: 'boolean',
-                hidden: true,
-                default: true,
-            })
-            .option('no-inject', {
-                type: 'boolean',
-                desc: 'Disallow injecting additional data from Rosetta',
-            })
-            .option('strict-fields', {
-                type: 'boolean',
-                hidden: true,
-                default: true,
-            })
-            .option('no-strict-fields', {
-                type: 'boolean',
-                desc: 'Marks classes as accepting fields of any type',
-            })
-            .option('rosetta', {
-                type: 'string',
-                alias: 'r',
-                desc: 'The directory to use for rosetta files',
-            })
-            .option('exclude', {
-                type: 'array',
-                alias: 'e',
-                string: true,
-                desc: 'Classes to exclude from annotations',
-            })
-            .option('exclude-fields', {
-                type: 'array',
-                string: true,
-                desc: 'Classes to include without fields',
-            })
-            .option('exclude-known-defs', {
-                type: 'boolean',
-                default: true,
-                desc: 'Whether known definition classes should be included without fields',
-            }),
-    )
+    yargs = sharedPrefix(yargs)
+        .option('output-directory', {
+            type: 'string',
+            alias: 'o',
+            required: true,
+            desc: 'The directory for output stubs',
+        })
+        .option('alphabetize', {
+            type: 'boolean',
+            default: true,
+            desc: 'Whether fields and functions should be alphabetically sorted',
+        })
+        .option('include-kahlua', {
+            type: 'boolean',
+            alias: 'k',
+            desc: 'Whether to generate the kahlua stub.',
+        })
+        .option('inject', {
+            type: 'boolean',
+            hidden: true,
+            default: true,
+        })
+        .option('no-inject', {
+            type: 'boolean',
+            desc: 'Disallow injecting additional data from Rosetta',
+        })
+        .option('strict-fields', {
+            type: 'boolean',
+            hidden: true,
+            default: true,
+        })
+        .option('no-strict-fields', {
+            type: 'boolean',
+            desc: 'Marks classes as accepting fields of any type',
+        })
+        .option('rosetta', {
+            type: 'string',
+            alias: 'r',
+            desc: 'The directory to use for rosetta files',
+        })
+
+    return sharedSuffix(addExcludeOptions(yargs))
 }
 
 /**
  * Adds the CLI options for the rosetta initialization command.
  */
 const initRosettaCommand = (yargs: yargs.Argv) => {
-    return sharedSuffix(
-        sharedPrefix(yargs)
-            .option('output-directory', {
-                type: 'string',
-                alias: 'o',
-                required: true,
-                desc: 'The directory for output stubs',
-            })
-            .option('format', {
-                type: 'string',
-                alias: 'f',
-                default: 'yml',
-                choices: ['json', 'yml'],
-                desc: 'The format to use for generated files',
-            })
-            .option('exclude', {
-                type: 'array',
-                alias: 'e',
-                string: true,
-                desc: 'Classes to exclude from annotations',
-            })
-            .option('exclude-fields', {
-                type: 'array',
-                string: true,
-                desc: 'Classes to include without fields',
-            })
-            .option('exclude-known-defs', {
-                type: 'boolean',
-                default: true,
-                desc: 'Whether known definition classes should be included without fields',
-            }),
-    )
+    yargs = sharedPrefix(yargs)
+        .option('output-directory', {
+            type: 'string',
+            alias: 'o',
+            required: true,
+            desc: 'The directory for output stubs',
+        })
+        .option('format', {
+            type: 'string',
+            alias: 'f',
+            default: 'yml',
+            choices: ['json', 'yml'],
+            desc: 'The format to use for generated files',
+        })
+
+    return sharedSuffix(addExcludeOptions(yargs))
 }
 
 /**
  * Adds the shared CLI options for report commands.
  */
 const reportCommand = (yargs: yargs.Argv) => {
-    return sharedSuffix(
-        sharedPrefix(yargs).option('output-file', {
-            type: 'string',
-            alias: 'o',
-            desc: 'The output file for report results',
-        }),
-    )
+    yargs = sharedPrefix(yargs).option('output-file', {
+        type: 'string',
+        alias: 'o',
+        desc: 'The output file for report results',
+    })
+
+    return sharedSuffix(yargs)
 }
 
 yargs(hideBin(process.argv))
@@ -187,10 +179,9 @@ yargs(hideBin(process.argv))
     )
     .command(
         'init-rosetta',
-        'Generates default Rosetta files',
+        'Generates default Rosetta data files',
         initRosettaCommand,
-        (async (args: AnnotateArgs) =>
-            await new Annotator(args).runRosettaInitialization()) as any,
+        (async (args: GenerateArgs) => await new Generator(args).run()) as any,
     )
     .command(
         'report-analysis',
