+++ src/rosetta/index.ts
@@ -1,2 +1,3 @@
 export { Rosetta } from './Rosetta'
+export { RosettaGenerator } from './RosettaGenerator'
 export type * from './types'
